refactor: separate the extension from runtime
