+++ Cargo.toml
@@ -24,18 +24,19 @@ members = [
     "primitives",
     "scripts/fund-dev-accounts",
 ]
-exclude = [
-    "pop-api",
-    "tests/contracts"
-]
+exclude = ["pop-api", "tests/contracts"]
 
 resolver = "2"
 
 [workspace.dependencies]
-codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
+codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = [
+    "derive",
+] }
 hex-literal = "0.4.1"
 log = { version = "0.4.20", default-features = false }
-scale-info = { version = "2.10.0", default-features = false, features = ["derive"] }
+scale-info = { version = "2.10.0", default-features = false, features = [
+    "derive",
+] }
 smallvec = "1.11.0"
 serde = "1.0.195"
 clap = { version = "4.4.18", features = ["derive"] }
@@ -53,9 +54,10 @@ substrate-build-script-utils = "11.0.0"
 
 # Local
 pallet-api = { path = "pallets/api", default-features = false }
-pop-runtime-devnet = { path = "runtime/devnet", default-features = true } # default-features=true required for `-p pop-node` builds
-pop-runtime-testnet = { path = "runtime/testnet", default-features = true } # default-features=true required for `-p pop-node` builds
+pop-runtime-devnet = { path = "runtime/devnet", default-features = true }          # default-features=true required for `-p pop-node` builds
+pop-runtime-testnet = { path = "runtime/testnet", default-features = true }        # default-features=true required for `-p pop-node` builds
 pop-runtime-common = { path = "runtime/common", default-features = false }
+pop-runtime-extensions = { path = "runtime/extensions", default-features = false }
 pop-primitives = { path = "./primitives", default-features = false }
 
 # Substrate
@@ -143,7 +145,9 @@ xcm-executor = { version = "8.0.2", package = "staging-xcm-executor", default-fe
 asset-hub-rococo-runtime = { version = "0.12.3", default-features = false }
 asset-test-utils = { version = "8.0.1", default-features = false }
 cumulus-pallet-aura-ext = { version = "0.8.0", default-features = false }
-cumulus-pallet-parachain-system = { version = "0.8.1", default-features = false, features = ["parameterized-consensus-hook"] }
+cumulus-pallet-parachain-system = { version = "0.8.1", default-features = false, features = [
+    "parameterized-consensus-hook",
+] }
 cumulus-pallet-session-benchmarking = { version = "10.0.0", default-features = false }
 cumulus-pallet-xcm = { version = "0.8.0", default-features = false }
 cumulus-pallet-xcmp-queue = { version = "0.8.0", default-features = false }
