+++ runtime/devnet/Cargo.toml
@@ -24,6 +24,7 @@ smallvec.workspace = true
 # Local
 pop-primitives.workspace = true
 pop-runtime-common.workspace = true
+pop-runtime-extensions.workspace = true
 pallet-api.workspace = true
 
 # Substrate
@@ -140,6 +141,7 @@ std = [
     "polkadot-parachain-primitives/std",
     "polkadot-runtime-common/std",
     "pop-primitives/std",
+    "pop-runtime-extensions/std",
     "scale-info/std",
     "sp-api/std",
     "sp-io/std",
