+++ runtime/extensions/Cargo.toml
@@ -1,0 +1,52 @@
+[package]
+name = "pop-runtime-extensions"
+version = "0.1.0"
+authors.workspace = true
+description.workspace = true
+license = "Unlicense"
+homepage.workspace = true
+repository.workspace = true
+edition.workspace = true
+publish = false
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+codec.workspace = true
+log.workspace = true
+
+# Local
+pop-primitives.workspace = true
+pop-runtime-common.workspace = true
+pallet-api.workspace = true
+
+# Substrate
+frame-support.workspace = true
+frame-system.workspace = true
+pallet-assets.workspace = true
+pallet-contracts.workspace = true
+sp-core.workspace = true
+sp-runtime.workspace = true
+sp-std.workspace = true
+
+[features]
+default = ["std"]
+std = [
+ "codec/std",
+ "frame-support/std",
+ "frame-system/std",
+ "log/std",
+ "pallet-assets/std",
+ "pallet-contracts/std",
+ "pallet-api/std",
+ "pop-primitives/std",
+ "sp-core/std",
+ "sp-std/std",
+]
+runtime-benchmarks = [
+ "frame-support/runtime-benchmarks",
+ "frame-system/runtime-benchmarks",
+ "pallet-assets/runtime-benchmarks",
+ "pallet-contracts/runtime-benchmarks",
+]
